+++ R/subBoot.R
@@ -23,6 +23,8 @@ thetaExtract <- function(merMod){
 #' @param FUN the function to apply to each bootstrapped model
 #' @param R the number of bootstrap replicates, default is 100
 #' @param seed numeric, optional argument to set seed for simulations
+#' @param warn logical, if TRUE, warnings from lmer will be issued, otherwise they will be suppressed
+#' default is FALSE
 #' @return a data.frame of parameters extracted from each of the R replications.
 #' The original values are appended to the top of the matrix.
 #' @details This function allows users to estimate parameters of a
@@ -33,9 +35,15 @@ thetaExtract <- function(merMod){
 #' resultMatrix <- subBoot(fm1, n = 160, FUN = thetaExtract, R = 20)
 #' }
 #' @export
-subBoot <- function(merMod, n = NULL, FUN, R = 100, seed=NULL){
-  if(missing(n))
+subBoot <- function(merMod, n = NULL, FUN, R = 100, seed = NULL, warn = FALSE){
+  if (missing(n)) {
     n <- nrow(merMod@frame)
+  }
+
+  if (!warn) {
+    message("Warnings set to off by default, not all submodels may have converged.")
+  }
+
   resultMat <- matrix(FUN(merMod), nrow = 1)
   tmp <- matrix(data=NA, nrow=R, ncol=ncol(resultMat))
   resultMat <- rbind(resultMat, tmp); rm(tmp)
@@ -51,7 +59,22 @@ subBoot <- function(merMod, n = NULL, FUN, R = 100, seed=NULL){
       # http://proteo.me.uk/2013/12/fast-subset-selection-by-row-name-in-r/
       newdata <- merMod@frame[match(mysamp, rownames(merMod@frame)),]
       # Only for lmerMod
-      tmpMod <- lmer(formula(merMod), data = newdata)
+      if (!warn) {
+        suppressWarnings({
+          if (class(merMod) == "lmerMod") {
+            tmpMod <- lmer(formula(merMod), data = newdata)
+          } else if (class(merMod) == "glmerMod") {
+            tmpMod <- glmer(formula(merMod), data = newdata, family = merMod@call$family)
+          }
+
+        })
+      } else {
+        if (class(merMod) == "lmerMod") {
+          tmpMod <- lmer(formula(merMod), data = newdata)
+        } else if (class(merMod) == "glmerMod") {
+          tmpMod <- glmer(formula(merMod), data = newdata, family = merMod@call$family)
+        }
+      }
       resultMat[i + 1, ] <- FUN(tmpMod)
     }
   resultMat <- data.frame(param=resultMat)