+++ NEWS.md
@@ -1,5 +1,22 @@
 # NEWS
 
+## merTools 0.5.0
+
+### New Features
+
+- `subBoot` now works with `glmerMod` objects as well
+- `reMargins` a new function that allows the user to marginalize the prediction over breaks in the 
+distribution of random effect distributions, see `?reMargins` and the new `reMargins` vignette (closes #73)
+
+### Bug fixes
+
+- Fixed an issue where known convergence errors were issuing warnings and causing the test suite 
+to not work
+- Fixed an issue where models with a random slope, no intercept, and no fixed term were unable 
+to be predicted (#101)
+- Fixed an issue with shinyMer not working with substantive fixed effects (#93)
+
+
 ## merTools 0.4.2
 
 ### New Features
@@ -15,7 +32,7 @@ in the `Suggests` field
 wrong dimensions when predicting a single row of observations for a `glm`
 - Fixed a bug (#96) related to `rstanarm` dependencies in the package vignette
 - Switched from `dontrun` to `donttest` for long-running examples (CRAN compliance)
-- Fixed and made more clear the generics applying to `merModList` objects
+- Fixed and made more clear the generics applying to `merModList` objects (#92)
 
 ## merTools 0.4.1